modified the data loader for blinks onset checks; modified blink stats to use a file for group names
the group_csv_data_loader file is fixed to calculate the onset blinks (fixed boolean variable);
the blinks_stats file contains the option of loading the group names by file, including the active and valid times within the group conversation.  This data is in a different file, along with each group name.  Currently the script is not fully working for the file with the group names options.